Update install scripts

+++ MDI_Mechanic/scripts/install_engine.py
@@ -1,14 +1,22 @@
 import os
-import sys
-import yaml
+import utils.utils as ut
 
-# Path to this file
-file_path = os.path.dirname(os.path.realpath(__file__))
-base_path = os.path.dirname( os.path.dirname( file_path ) )
+# Path to the user directory
+base_path = ut.get_base_path()
 
-# Build the docker image
-#os.system("docker build -t mdi_mechanic/lammps user")
+# Switch to the base directory
+os.chdir(base_path)
 
-# Run the engine, using Docker
+# Build the MDI Mechanic image
+ret = os.system("docker build -t mdi_mechanic/mdi_mechanic MDI_Mechanic/docker")
+if ret != 0:
+    raise Exception("Unable to build the MDI Mechanic image")
+
+# Build the engine image
+ret = os.system("docker build -t mdi_mechanic/lammps user")
+if ret != 0:
+    raise Exception("Unable to build the engine image")
+
+# Build the engine, within its Docker image
 docker_string = "docker run --rm -v " + str(base_path) + ":/repo -it mdi_mechanic/lammps bash -c \"cd /repo/user && ls && ./docker_install.sh\""
 os.system(docker_string)
